signals be foken hard

+++ src/utils/signal_handler.c
@@ -10,7 +10,7 @@ void	init_child_signals(void)
 
 	ft_bzero(&sa, sizeof(sa));
 	sa.sa_sigaction = sig_handler;
-	sa.sa_flags = SA_SIGINFO;
+	sa.sa_flags = SA_RESTART;
 	sigaction(SIGINT, &sa, NULL);
 }
 
@@ -24,7 +24,7 @@ void	init_signals(void)
 	sa.sa_sigaction = sig_handler;
 	sa.sa_flags = SA_SIGINFO;
 	ignore.sa_handler = SIG_IGN;
-	sigaction(SIGINT, &ignore, NULL);
+	sigaction(SIGINT, &sa, NULL);
 	sigaction(SIGQUIT, &ignore, NULL);
 }
 
@@ -32,11 +32,12 @@ void	sig_handler(int sig, siginfo_t *info, void *ucontent)
 {
 	(void)ucontent;
 	(void)info;
-	if (sig == SIGINT)
+	if (sig == SIGINT && info->si_pid != 0)
 	{
 		//r_sig = sig;
+		//printf("sig pid: %d\n", info->si_pid);
 		write(1, "\n", 1);
-		 kill (info->si_pid, SIGTERM);
+		kill (info->si_pid, SIGTERM);
 	}
 }
